zhipu-ai-config/requirements.txt
# 核心依赖
cryptography>=41.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
pyyaml>=6.0

# 开发和测试依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
ruff>=0.1.0
mypy>=1.5.0

# 可选依赖
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.5.0